old_requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
cassandra-driver==3.29.1
cassio==0.1.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ctransformers==0.2.27
dataclasses-json==0.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
exceptiongroup==1.2.0
fastjsonschema==2.19.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
gdown==5.1.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
gradientai==1.11.0
greenlet==3.0.3
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lazy_loader==0.4
librosa==0.10.1
llama-index==0.10.30
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.12
llama-index-core==0.10.30
llama-index-embeddings-gradient==0.1.2
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-gradient==0.1.2
llama-index-llms-openai==0.1.15
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-cassandra==0.1.3
llama-index-vector-stores-postgres==0.1.5
llama-parse==0.4.1
llamaindex-py-client==0.1.18
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
openai==1.21.2
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
pgvector==0.2.5
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyarrow==15.0.2
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
pypdf==4.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.0
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.3
SQLAlchemy==2.0.29
streamlit==1.33.0
streamlit-image-select==0.6.0
streamlit_mic_recorder==0.0.8
striprtf==0.0.26
sympy==1.12
tbb==2021.12.0
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
webencodings==0.5.1
wrapt==1.16.0
yarl==1.9.4
